Consolidate docs
- Rework README
- Convert tutorial from HTML to Markdown
- Convert website content to Markdown
- Convert overview to Markdown
- Rework documention structure
- Add general & setup documentation
