+++ GNUmakefile
@@ -35,7 +35,7 @@ BINDIR = bin
 SRCDIR = sljit_src
 TESTDIR = test_src
 REGEXDIR = regex_src
-EXAMPLEDIR = docs/tutorial
+EXAMPLEDIR = docs/tutorial/sources
 
 TARGET = $(BINDIR)/sljit_test $(BINDIR)/regex_test
 EXAMPLE_TARGET = $(BINDIR)/func_call $(BINDIR)/first_program $(BINDIR)/branch $(BINDIR)/loop $(BINDIR)/array_access $(BINDIR)/func_call $(BINDIR)/struct_access $(BINDIR)/temp_var $(BINDIR)/brainfuck
