+++ docs/README.md
@@ -1,0 +1,3 @@
+# Documentation
+
+This folder contains a series of subfolders ([`general`](/docs/general/introduction.md) and [`tutorial`](/docs/tutorial/01-overview.md)) with documentation in [Markdown](https://en.wikipedia.org/wiki/Markdown).
