+++ docs/general/contributing.md
@@ -1,0 +1,9 @@
+# Contributing
+
+SLJIT needs your help! You can contribute by:
+- Reporting bugs [here](https://github.com/zherczeg/sljit/issues)
+- Submitting patches [here](https://github.com/zherczeg/sljit/pulls)
+- Improving SLJIT's documentation
+- Testing SLJIT on all sorts of platforms and compilers (especially the not so common ones like MIPS, PowerPC and SPARC)
+
+Thank you!
