+++ docs/tutorial/01-overview.md
@@ -1,0 +1,12 @@
+# Overview
+
+> [!NOTE]
+> The original version of this tutorial was provided by Wen Xichang (wenxichang@163.com) in 2015.
+
+To be able to follow this tutorial, you will need:
+- A `C99` compatible C/C++ compiler (somewhat recent versions of GCC / Clang / MSVC are safe choices)
+- GNU Make (or CMake, if you are on Windows)
+
+The examples shown in the tutorial can be built via the `examples` make target. See the [setup guide](/docs/general/getting-started/setup) for more details.
+
+The tutorial assumes you are already familiar with the C programming language.
