--- docs/tutorial/README
@@ -1,3 +1,0 @@
-These files were contributed by Wen Xichang.
-
-Copyright 2015 Wen Xichang (wenxichang@163.com). All rights reserved.
